pyproject.toml
[tool.poetry]
name = "bmp388"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
rpi-gpio = "^0.7.1"
spidev = "^3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
